+++ pages/_app.tsx
@@ -3,7 +3,7 @@ import { SessionProvider } from 'next-auth/react'
 import { ChakraProvider } from '@chakra-ui/react'
 
 import { theme } from '~/assets/theme'
-import { SidebarDrawerProvider } from '~/contexts'
+import { SidebarDrawerProvider } from '~/contexts/sidebar-drawer'
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
