--- src/components/index.ts
@@ -1,3 +1,0 @@
-export { Logo } from './logo'
-export * from './header'
-export * from './sidebar'
