+++ src/components/page-loader.tsx
@@ -1,0 +1,9 @@
+import { Center, Spinner } from '@chakra-ui/react'
+
+export function PageLoader() {
+  return (
+    <Center h="100vh">
+      <Spinner />
+    </Center>
+  )
+}
