+++ src/components/sidebar/active-link.tsx
@@ -1,6 +1,6 @@
+import { cloneElement, ReactElement } from 'react'
 import Link, { LinkProps } from 'next/link'
 import { useRouter } from 'next/router'
-import { cloneElement, ReactElement } from 'react'
 
 interface ActiveLinkProps extends LinkProps {
   children: ReactElement
